refactor(web): #127 admin-users 역할 유틸 공통화
PR #128 비블로커 코멘트 반영으로 role 관련 중복 정의를 공통 모듈로 추출했습니다.

- normalizeRoles/roleLabel/RoleChecklist를 role-shared.tsx로 이동

- page.tsx/create-form.tsx는 공통 모듈 import로 통일

- 기능 동작은 동일하며 유지보수 시 불일치 리스크를 줄였습니다.

Log: 2026-02-18 22:00 | Ji, hye-jin | refactor | admin-users 역할 유틸 공통화 | apps/web/app/admin/admin-users/role-shared.tsx,apps/web/app/admin/admin-users/page.tsx,apps/web/app/admin/admin-users/create-form.tsx,docs/dev_log_260218.md
