+++ apps/web/app/admin/admin-users/create-form.tsx
@@ -1,34 +1,8 @@
 "use client";
 
 import { useEffect, useState, type FormEvent } from 'react';
-import { ADMIN_PERMISSION_TOKENS } from '../../../lib/rbac';
 import { type AdminUserCreatePayload } from '../../../services/admin-users';
-
-const BASE_ROLE_TOKENS = ['member', 'admin', 'super_admin'] as const;
-const KNOWN_ROLE_TOKENS = [...BASE_ROLE_TOKENS, ...ADMIN_PERMISSION_TOKENS];
-
-function normalizeRoles(roles: string[]): string[] {
-  return Array.from(
-    new Set(
-      roles
-        .filter((role) => role.trim().length > 0)
-        .map((role) => role.trim().toLowerCase())
-    )
-  );
-}
-
-function roleLabel(role: string): string {
-  if (role === 'member') return 'member';
-  if (role === 'admin') return 'admin';
-  if (role === 'super_admin') return 'super_admin';
-  if (role === 'admin_posts') return '게시물 관리';
-  if (role === 'admin_events') return '행사 관리';
-  if (role === 'admin_hero') return '홈 배너 관리';
-  if (role === 'admin_notifications') return '알림 관리';
-  if (role === 'admin_signup') return '가입신청 심사';
-  if (role === 'admin_roles') return '권한 관리';
-  return role;
-}
+import { RoleChecklist, normalizeRoles } from './role-shared';
 
 function parseCohortValue(rawValue: string): number | null {
   const parsed = Number(rawValue);
@@ -39,35 +13,6 @@ function parseCohortValue(rawValue: string): number | null {
 
 function hasAdminGradeRole(roles: string[]): boolean {
   return roles.includes('admin') || roles.includes('super_admin');
-}
-
-function RoleChecklist({
-  studentId,
-  draftRoles,
-  disabled,
-  onToggle,
-}: {
-  studentId: string;
-  draftRoles: string[];
-  disabled: boolean;
-  onToggle: (studentId: string, role: string, checked: boolean) => void;
-}) {
-  return (
-    <div className="grid grid-cols-2 gap-1">
-      {KNOWN_ROLE_TOKENS.map((role) => (
-        <label key={`${studentId}:${role}`} className="text-xs text-text-primary">
-          <input
-            type="checkbox"
-            className="mr-1 align-middle"
-            checked={draftRoles.includes(role)}
-            disabled={disabled}
-            onChange={(e) => onToggle(studentId, role, e.currentTarget.checked)}
-          />
-          {roleLabel(role)}
-        </label>
-      ))}
-    </div>
-  );
 }
 
 function AdminGradeHint({ hasAdminGrade }: { hasAdminGrade: boolean }) {
