+++ apps/web/app/admin/admin-users/page.tsx
@@ -8,7 +8,7 @@ import { useToast } from '../../../components/toast';
 import { useAuth } from '../../../hooks/useAuth';
 import { ApiError } from '../../../lib/api';
 import { apiErrorToMessage } from '../../../lib/error-map';
-import { ADMIN_PERMISSION_TOKENS, isSuperAdminSession } from '../../../lib/rbac';
+import { isSuperAdminSession } from '../../../lib/rbac';
 import {
   createAdminUser,
   listAdminUsers,
@@ -16,44 +16,18 @@ import {
   type AdminUserCreatePayload,
   type AdminUserRolesRead,
 } from '../../../services/admin-users';
-
-const BASE_ROLE_TOKENS = ['member', 'admin', 'super_admin'] as const;
-const KNOWN_ROLE_TOKENS = [...BASE_ROLE_TOKENS, ...ADMIN_PERMISSION_TOKENS];
-const KNOWN_ROLE_SET = new Set<string>(KNOWN_ROLE_TOKENS);
+import { KNOWN_ROLE_SET, RoleChecklist, normalizeRoles } from './role-shared';
 
 type Feedback = {
   tone: 'success' | 'error';
   message: string;
 };
 
-function normalizeRoles(roles: string[]): string[] {
-  return Array.from(
-    new Set(
-      roles
-        .filter((role) => role.trim().length > 0)
-        .map((role) => role.trim().toLowerCase())
-    )
-  );
-}
-
 function isSameRoleSet(left: string[], right: string[]): boolean {
   const a = normalizeRoles(left).sort();
   const b = normalizeRoles(right).sort();
   if (a.length !== b.length) return false;
   return a.every((value, index) => value === b[index]);
-}
-
-function roleLabel(role: string): string {
-  if (role === 'member') return 'member';
-  if (role === 'admin') return 'admin';
-  if (role === 'super_admin') return 'super_admin';
-  if (role === 'admin_posts') return '게시물 관리';
-  if (role === 'admin_events') return '행사 관리';
-  if (role === 'admin_hero') return '홈 배너 관리';
-  if (role === 'admin_notifications') return '알림 관리';
-  if (role === 'admin_signup') return '가입신청 심사';
-  if (role === 'admin_roles') return '권한 관리';
-  return role;
 }
 
 function FeedbackBanner({ feedback }: { feedback: Feedback | null }) {
@@ -67,35 +41,6 @@ function FeedbackBanner({ feedback }: { feedback: Feedback | null }) {
   return (
     <div className={`rounded border px-3 py-2 text-sm ${className}`} role="status">
       {feedback.message}
-    </div>
-  );
-}
-
-function RoleChecklist({
-  studentId,
-  draftRoles,
-  disabled,
-  onToggle,
-}: {
-  studentId: string;
-  draftRoles: string[];
-  disabled: boolean;
-  onToggle: (studentId: string, role: string, checked: boolean) => void;
-}) {
-  return (
-    <div className="grid grid-cols-2 gap-1">
-      {KNOWN_ROLE_TOKENS.map((role) => (
-        <label key={`${studentId}:${role}`} className="text-xs text-text-primary">
-          <input
-            type="checkbox"
-            className="mr-1 align-middle"
-            checked={draftRoles.includes(role)}
-            disabled={disabled}
-            onChange={(e) => onToggle(studentId, role, e.currentTarget.checked)}
-          />
-          {roleLabel(role)}
-        </label>
-      ))}
     </div>
   );
 }
