+++ docs/dev_log_260218.md
@@ -47,3 +47,4 @@
 - Feat: `/admin/admin-users`에 super_admin 전용 관리자 생성 폼 추가(학번/이메일/이름/기수/임시비밀번호/초기권한) 및 생성 후 목록 재조회 연동
 - Test: `tests/api/test_admin_roles_api.py`에 관리자 생성 성공/권한없음(403)/중복충돌(409) 케이스 추가(7 passed)
 - Fix: PR #128 리뷰 반영 — 생성 성공 후 폼 리셋, roles 타입 `list[str]` 고정, 이메일 `type=email`/관리자등급 최소선택/cohort 숫자 검증 보강
+- Refactor: PR #128 비블로커 반영 — `admin-users`의 역할 유틸/체크리스트 중복을 `role-shared.tsx`로 공통화해 `page.tsx`/`create-form.tsx` 동기화 리스크 축소
